requirements.txt
numpy>=2.3.0
matplotlib>=3.10.3
pandas>=2.3.0
Pillow>=11.2.1
opencv-python>=4.11.0
scipy>=1.15.3
IPython>=9.3.0
notebook>=7.4.3